% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidthclass.R
\docType{class}
\name{bandwidth-class}
\alias{bandwidth}
\alias{bandwidth-class}
\title{An S4 class to represent a bandwidth for gw-xtabs eyc.}
\description{
An S4 class to represent a bandwidth for gw-xtabs eyc.
}
\section{Slots}{

\describe{
\item{\code{value}}{The value of the bandwidth}

\item{\code{units}}{The units of length used to specify a non-adaptive bandwidth}

\item{\code{adaptive}}{A logical value indicating whether bandwidth is adaptive}
}}

