% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{data}
\name{classification}
\alias{classification}
\title{Spatially referenced land use data}
\description{
A SpatialPointsDataFrame named \code{loc_confuse} containing actual (\code{actual}) and predicted (\code{svmpred}) land use at
a number of data points. It has the following columns:
\tabular{rl}{
  \code{actual} \tab Actual land use class \cr
  \code{svmpred} \tab Predicted land use class using a support vector machine approach
}
}
\keyword{data}

