% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{coordinates,SpatialCrossTabs-method}
\alias{coordinates,SpatialCrossTabs-method}
\title{Return coordinates of a \code{SpatialCrossTabs} object}
\usage{
\S4method{coordinates}{SpatialCrossTabs}(obj, ...)
}
\arguments{
\item{obj}{A \code{SpatialCrossTabs} object}
}
\value{
The coordinates as a \code{c(n,2)} matrix
}
\description{
Return coordinates of a \code{SpatialCrossTabs} object
}

