% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidthclass.R
\name{fixed}
\alias{fixed}
\title{Fixed bandwidth}
\usage{
fixed(n, units = "km")
}
\arguments{
\item{n}{- value (distance) of bandwidth}

\item{units}{- Distance units used ('km','m','mi','deg','rad','nm' or NA)}
}
\description{
Create a fixed bandwidth object
}

