% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{data}
\name{gcr2km}
\alias{gcr2km}
\title{gcr2km: Great circle radian angle to kilometers conversion}
\format{An object of class \code{numeric} of length 1.}
\usage{
gcr2km

gcr2km
}
\description{
gcr2km: Great circle radian angle to kilometers conversion

gcr2km: Great circle radian angle to kilometers conversion
}
\keyword{datasets}

