% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{data}
\name{gcr2mi}
\alias{gcr2mi}
\title{gcr2mi: Great circle radian angle to miles conversion}
\format{An object of class \code{numeric} of length 1.}
\usage{
gcr2mi
}
\description{
gcr2mi: Great circle radian angle to miles conversion
}
\keyword{datasets}

