% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{data}
\name{gcr2nm}
\alias{gcr2nm}
\title{gcr2nm: Great circle radian angle to nautical miles conversion}
\format{An object of class \code{numeric} of length 1.}
\usage{
gcr2nm
}
\description{
gcr2nm: Great circle radian angle to nautical miles conversion
}
\keyword{datasets}

