% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{package}
\name{gwxtab}
\alias{gwxtab}
\alias{gwxtab,SpatialCrossTabs,SpatialPoints,numeric-method}
\alias{gwxtab,SpatialCrossTabs,missing,numeric-method}
\alias{gwxtab-package}
\title{gwxtab: A package for manipulating geographically weighted crosstabulations}
\usage{
gwxtab(spxt, xtab_pts, bw)

\S4method{gwxtab}{SpatialCrossTabs,missing,numeric}(spxt, bw)

\S4method{gwxtab}{SpatialCrossTabs,SpatialPoints,numeric}(spxt, xtab_pts, bw)
}
\arguments{
\item{spxt}{A SpatialCrossTabs object}

\item{xtab_pts}{A \code{SpatialPoints} object (optional - if omitted points are as in \code{spxt})}

\item{bw}{Bandwidth}
}
\value{
Weighted sum of all contingency tables in \code{spxt}
}
\description{
For more details see the help vignette:
\code{vignette("basic_use", package = "gwxtab")}

Geographically Weighted contingency tables
}

