% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{gwxtab_probe}
\alias{gwxtab_probe}
\alias{gwxtab_probe,SpatialCrossTabs,bandwidth,character,function-method}
\alias{gwxtab_probe,SpatialCrossTabs,bandwidth,missing,function-method}
\alias{gwxtab_probe,SpatialCrossTabs,bandwidth,missing,missing-method}
\alias{gwxtab_probe,SpatialCrossTabs,numeric,character,function-method}
\alias{gwxtab_probe,SpatialCrossTabs,numeric,character,missing-method}
\alias{gwxtab_probe,SpatialCrossTabs,numeric,missing,missing-method}
\title{Geographically Weighted contingency table probing functions}
\usage{
gwxtab_probe(spxt, bw, mode, melt)

\S4method{gwxtab_probe}{SpatialCrossTabs,numeric,missing,missing}(spxt, bw)

\S4method{gwxtab_probe}{SpatialCrossTabs,bandwidth,missing,missing}(spxt, bw)

\S4method{gwxtab_probe}{SpatialCrossTabs,bandwidth,missing,`function`}(spxt, bw,
  melt)

\S4method{gwxtab_probe}{SpatialCrossTabs,numeric,character,missing}(spxt, bw,
  mode)

\S4method{gwxtab_probe}{SpatialCrossTabs,numeric,character,`function`}(spxt, bw,
  mode, melt)

\S4method{gwxtab_probe}{SpatialCrossTabs,bandwidth,character,`function`}(spxt,
  bw, mode, melt)
}
\arguments{
\item{spxt}{A SpatialCrossTabs object}

\item{bw}{Bandwidth}

\item{mode}{Kind of parameters to give to probe: \code{scalar} for \code{(x,y)} pair, \code{vector} for \code{X} length 2 vector, \code{matrix} for \code{c(n,2)} matrix, or \code{Spatial} for a a spatial object}

\item{melt}{'melting' operator}
}
\value{
Function to return geographically weighted sum of all contingency tables in \code{spxt} centred on probe argument.
}
\description{
Geographically Weighted contingency table probing functions
}

