% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{gwxtab_sample}
\alias{gwxtab_sample}
\alias{gwxtab_sample,Raster,SpatialCrossTabs,bandwidth,function-method}
\alias{gwxtab_sample,Spatial,SpatialCrossTabs,bandwidth,function-method}
\alias{gwxtab_sample,Spatial,SpatialCrossTabs,numeric,function-method}
\title{Geographically Weighted contingency table sampling functions}
\usage{
gwxtab_sample(spdf, spxt, bw, melt)

\S4method{gwxtab_sample}{Spatial,SpatialCrossTabs,numeric,`function`}(spdf,
  spxt, bw, melt)

\S4method{gwxtab_sample}{Spatial,SpatialCrossTabs,bandwidth,`function`}(spdf,
  spxt, bw, melt)

\S4method{gwxtab_sample}{Raster,SpatialCrossTabs,bandwidth,`function`}(spdf,
  spxt, bw, melt)
}
\arguments{
\item{spxt}{A SpatialCrossTabs object}

\item{bw}{Bandwidth}

\item{melt}{'melting' operator}
}
\value{
Function to return geographically weighted sum of all contingency tables in \code{spxt} centred on probe argument.
}
\description{
Geographically Weighted contingency table sampling functions
}

