% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{is.projected,Raster-method}
\alias{is.projected,Raster-method}
\title{Check whether Raster object uses projected  coordinates}
\usage{
\S4method{is.projected}{Raster}(obj)
}
\description{
This is exactly the same as the \code{is.projected} method for \code{Spatial} objects
}

