% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{new_spxt}
\alias{new_spxt}
\alias{new_spxt,SpatialPointsDataFrame,character,character-method}
\alias{new_spxt,SpatialPointsDataFrame,table,missing-method}
\alias{new_spxt,matrix,table,missing-method}
\title{Create a new `SpatialCrossTabs` object}
\usage{
new_spxt(spdf, var1, var2)

\S4method{new_spxt}{SpatialPointsDataFrame,table,missing}(spdf, var1)

\S4method{new_spxt}{matrix,table,missing}(spdf, var1)

\S4method{new_spxt}{SpatialPointsDataFrame,character,character}(spdf, var1,
  var2)
}
\arguments{
\item{spdf}{A spatial object or nx2 matrix of coordinates}

\item{var1}{A 2-dimensional table object,  or a string specifying row variable}

\item{var2}{A string specifying column variable (if \code{var1} specifies row variable)}
}
\description{
Create a new `SpatialCrossTabs` object
}

