% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{proj4string,SpatialCrossTabs-method}
\alias{proj4string,SpatialCrossTabs-method}
\title{Return \code{proj4} string for map of projection}
\usage{
\S4method{proj4string}{SpatialCrossTabs}(obj)
}
\description{
Return \code{proj4} string for map of projection
}

