% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{projection_transform}
\alias{projection_transform}
\alias{spTransform,SpatialCrossTabs,CRS-method}
\alias{spTransform,SpatialCrossTabs,Spatial-method}
\alias{spTransform,SpatialCrossTabs,character-method}
\title{spTransform method for \code{SpatialCrossTabs}}
\usage{
\S4method{spTransform}{SpatialCrossTabs,CRS}(x, CRSobj)

\S4method{spTransform}{SpatialCrossTabs,character}(x, CRSobj)

\S4method{spTransform}{SpatialCrossTabs,Spatial}(x, CRSobj)
}
\arguments{
\item{x}{The \code{SpatialCrossTabs} object.}

\item{CRSobj}{Either a \code{CRS} object, a \code{PROJ4} string or a \code{Spatial} object to extract the \code{PROJ4} string from.}
}
\value{
A \code{SpatialCrossTabs} object with transformed coordinates.
}
\description{
Changes the map projection for a \code{SpatialCrossTabs} object.
}

