% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{set_proj4string}
\alias{proj4string<-,SpatialCrossTabs,CRS-method}
\alias{proj4string<-,SpatialCrossTabs,character-method}
\alias{set_proj4string}
\title{Set \code{proj4} string for map of projection}
\usage{
\S4method{proj4string}{SpatialCrossTabs,CRS}(obj) <- value

\S4method{proj4string}{SpatialCrossTabs,character}(obj) <- value
}
\description{
Set \code{proj4} string for map of projection
}

