% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{show,SpatialCrossTabs-method}
\alias{show,SpatialCrossTabs-method}
\title{Print out basic (short form) information for a SpatialCrossTabs object}
\usage{
\S4method{show}{SpatialCrossTabs}(object)
}
\arguments{
\item{object}{A SpatialCrossTabs object}
}
\description{
Print out basic (short form) information for a SpatialCrossTabs object
}
\note{
Not usually called explicitly
}

