% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{spTransform,SpatialCrossTabs,CRS-method}
\alias{spTransform,SpatialCrossTabs,CRS-method}
\title{spTransform method for \code{SpatialCrossTabs}}
\usage{
\S4method{spTransform}{SpatialCrossTabs,CRS}(x, CRSobj)
}
\description{
spTransform method for \code{SpatialCrossTabs}
}

