% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{[,SpatialCrossTabs,numeric,ANY-method}
\alias{[,SpatialCrossTabs,numeric,ANY-method}
\title{Access the nth crosstabulation in a SpatialCrossTabs object}
\usage{
\S4method{[}{SpatialCrossTabs,numeric,ANY}(x, i)
}
\arguments{
\item{x}{A SpatialCrossTabs object}

\item{i}{the index of the crosstabulation to extract}
}
\value{
The \code{i}th crosstabulation in \code{x}.
}
\description{
Access the nth crosstabulation in a SpatialCrossTabs object
}

