% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{[<-,SpatialCrossTabs,numeric,missing,table-method}
\alias{[<-,SpatialCrossTabs,numeric,missing,table-method}
\title{Assign the nth crosstabulation in a SpatialCrossTabs object}
\usage{
\S4method{[}{SpatialCrossTabs,numeric,missing,table}(x, i) <- value
}
\arguments{
\item{x}{A SpatialCrossTabs object}

\item{i}{the index of the crosstabulation to extract}

\item{A}{contingency table}
}
\description{
Assign the nth crosstabulation in a SpatialCrossTabs object
}

