% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{tabsum}
\alias{tabsum}
\alias{tabsum,SpatialCrossTabs-method}
\title{Sum of contingency tables}
\usage{
tabsum(spxt)

\S4method{tabsum}{SpatialCrossTabs}(spxt)
}
\arguments{
\item{spxt}{A SpatialCrossTabs object}
}
\value{
sum of all contingency tables in \code{spxt}
}
\description{
Sum of contingency tables
}

