% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCrossTabs.R
\docType{methods}
\name{wtabsum}
\alias{wtabsum}
\alias{wtabsum,SpatialCrossTabs,missing-method}
\alias{wtabsum,SpatialCrossTabs,numeric-method}
\title{Weighted sum of contingency tables}
\usage{
wtabsum(spxt, w)

\S4method{wtabsum}{SpatialCrossTabs,numeric}(spxt, w)

\S4method{wtabsum}{SpatialCrossTabs,missing}(spxt)
}
\arguments{
\item{spxt}{A SpatialCrossTabs object}

\item{w}{A set of weights (optional - default is \code{1/n})}
}
\value{
Weighted sum of all contingency tables in \code{spxt}
}
\description{
Weighted sum of contingency tables
}

